requirements.app.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6

pydantic==2.9.2
pydantic-settings==2.5.2

sqlalchemy==2.0.35
alembic==1.14.0
psycopg==3.2.3

redis==5.0.8

prometheus-client==0.20.0
requests==2.32.3
PyJWT[crypto]==2.10.1
numpy==2.4.1
av==16.0.1
jinja2==3.1.4
python-multipart==0.0.22
opentelemetry-api==1.29.0
opentelemetry-sdk==1.29.0
opentelemetry-exporter-otlp-proto-http==1.29.0
